Extract utility methods for Maven Extension and Gradle Plugin
- Remove Maven Extension usage of Guava
- Extract common utility methods into `Utils` class

This change simplifies the implementations CustomBuildScanConfig.
