+++ common-custom-user-data-gradle-plugin/src/main/java/com/gradle/Utils.java
@@ -1,0 +1,157 @@
+package com.gradle;
+
+import java.io.BufferedReader;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.Reader;
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.util.Properties;
+import java.util.concurrent.TimeUnit;
+
+class Utils {
+    private static final MessageDigest SHA_256 = createMessageDigest();
+
+    static boolean isNullOrEmpty(String value) {
+        return value == null || value.isEmpty();
+    }
+
+    static String sysProperty(String name) {
+        return System.getProperty(name);
+    }
+
+    static boolean sysPropertyPresent(String name) {
+        return !isNullOrEmpty(sysProperty(name));
+    }
+
+    static boolean sysPropertyKeyStartingWith(String keyPrefix) {
+        for (Object key : System.getProperties().keySet()) {
+            if (key instanceof String) {
+                String stringKey = (String) key;
+                if (stringKey.startsWith(keyPrefix)) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
+    static String envVariable(String name) {
+        return System.getenv(name);
+    }
+
+    static boolean envVariablePresent(String name) {
+        return !isNullOrEmpty(envVariable(name));
+    }
+
+    static String appendIfMissing(String str, String suffix) {
+        return str.endsWith(suffix) ? str : str + suffix;
+    }
+
+    static String urlEncode(String str) {
+        try {
+            return URLEncoder.encode(str, StandardCharsets.UTF_8.name());
+        } catch (UnsupportedEncodingException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    static Properties readPropertiesFile(String name) {
+        try (InputStream input = new FileInputStream(name)) {
+            Properties properties = new Properties();
+            properties.load(input);
+            return properties;
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    static boolean execAndCheckSuccess(String... args) {
+        Runtime runtime = Runtime.getRuntime();
+        Process process = null;
+        try {
+            process = runtime.exec(args);
+            boolean finished = process.waitFor(10, TimeUnit.SECONDS);
+            return finished && process.exitValue() == 0;
+        } catch (IOException | InterruptedException ignored) {
+            return false;
+        } finally {
+            if (process != null) {
+                process.destroyForcibly();
+            }
+        }
+    }
+
+    @SuppressWarnings("UnstableApiUsage")
+    static String execAndGetStdOut(String... args) {
+        Runtime runtime = Runtime.getRuntime();
+        Process process;
+        try {
+            process = runtime.exec(args);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        try (Reader standard = new BufferedReader(new InputStreamReader(process.getInputStream(), Charset.defaultCharset()))) {
+            try (Reader error = new BufferedReader(new InputStreamReader(process.getErrorStream(), Charset.defaultCharset()))) {
+                String standardText = readFully(standard);
+                String ignore = readFully(error);
+
+                boolean finished = process.waitFor(10, TimeUnit.SECONDS);
+                return finished && process.exitValue() == 0 ? trimAtEnd(standardText) : null;
+            }
+        } catch (IOException | InterruptedException e) {
+            throw new RuntimeException(e);
+        } finally {
+            process.destroyForcibly();
+        }
+    }
+
+    private static String readFully(Reader reader) throws IOException {
+        StringBuilder sb = new StringBuilder();
+        char[] buf = new char[1024];
+        int nRead;
+        while ((nRead = reader.read(buf)) != -1) {
+            sb.append(buf, 0, nRead);
+        }
+        return sb.toString();
+    }
+
+    private static String trimAtEnd(String str) {
+        return ('x' + str).trim().substring(1);
+    }
+
+    static String hashValue(Object value) {
+        if (value == null) {
+            return null;
+        }
+        String string = String.valueOf(value);
+        byte[] encodedHash = SHA_256.digest(string.getBytes());
+        StringBuilder hexString = new StringBuilder();
+        for (int i = 0; i < encodedHash.length / 4; i++) {
+            String hex = Integer.toHexString(0xff & encodedHash[i]);
+            if (hex.length() == 1) {
+                hexString.append("0");
+            }
+            hexString.append(hex);
+        }
+        hexString.append("...");
+        return hexString.toString();
+    }
+
+    private static MessageDigest createMessageDigest() {
+        try {
+            return MessageDigest.getInstance("SHA-256");
+        } catch (NoSuchAlgorithmException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private Utils() {}
+}
